refactor: 쿠키설정 관련 함수들 utils로 로직 분리하였음
